namespace abstraction for getVotingPower
